requirements.txt
line-bot-sdk
flask
openai
